feat: boost lint
